requirements/base.txt
Django>=4.2,<5.0
django-tenants>=3.5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-filter>=23.3
django-environ>=0.10
django-redis>=5.3
channels>=4.1
channels-redis>=4.2
daphne>=4.1
psycopg[binary]>=3.1
celery>=5.3
redis>=4.5
